pyproject.toml
[tool.poetry]
name = "orr-cli"
version = "0.1.40"
description = ""
authors = ["Jairo Matos da Rocha <devjairomr@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
orr = "orr_cli.cli:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.15.1"
rich = "^13.9.4"
psycopg2-binary = "^2.9.10"
sqlmodel = "^0.0.22"
geoalchemy2 = "^0.17.0"
dynaconf = "^3.2.7"
geopandas = "^1.0.1"
unidecode = "^1.3.7"
babel = "^2.14.0"
matplotlib = "^3.10.0"
folium = "^0.19.4"
loguru = "^0.7.3"
mkdocs-static-i18n = {extras = ["material"], version = "^1.3.0"}
nltk = "==3.8.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
blue = "^0.9.1"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"
mkdocs-macros-plugin = "^1.0.5"


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.4.14"
mkdocstrings = "^0.24.0"
mkdocstrings-python = "^1.7.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
docs = "python docs.py && mkdocs serve -a localhost:8200"
pre_test = "task lint"
test = "pytest -s -x --cov=orr_cli -vv"
post_test = "coverage html"
